Restricted creating, editing and deleting posts to those belonging to logged in user

+++ controllers/postController.js
@@ -16,20 +16,28 @@ module.exports.posts_get = async(req, res) => {
 
 module.exports.post_create_get = (req, res) => {
     const loggedUser = res.locals.user;
+    const { username } = req.params;
 
-    res.render("addPost", { loggedUser: loggedUser })
+    if (loggedUser.username === username) {
+        res.render("addPost", { loggedUser: loggedUser })
+    } else {
+        res.send("You are not allowed to access this route");
+    }
+    
 }
 
 module.exports.post_create_post = async (req, res) => {
     const loggedUser = res.locals.user;
     const { title, body } = req.body;
 
-    try {
-        const user = await User.findOne({ username: loggedUser.username });
-        const post = await Post.create({ title: title, body: body, author: user._id });
-        res.json({ post });
-    } catch (err){
-        console.log(err);
+    if (loggedUser.username === username) {
+        try {
+            const user = await User.findOne({ username: loggedUser.username });
+            const post = await Post.create({ title: title, body: body, author: user._id });
+            res.json({ post });
+        } catch (err){
+            console.log(err);
+        }
     }
 }
 
@@ -57,25 +65,35 @@ module.exports.post_detail = async(req, res) => {
 
 module.exports.post_edit_get = async(req, res) => {
     const loggedUser = res.locals.user;
-    const { slug } = req.params;
+    const { username, slug } = req.params;
 
-    try {
-        const post = await (await Post.findOne({ slug })).populate("author");
-        res.render("postEdit", { loggedUser: loggedUser, post: post });
-    } catch(err) {
-        console.log(err);
+    if (loggedUser.username === username) {
+        try {
+            const post = await (await Post.findOne({ slug })).populate("author");
+            res.render("postEdit", { loggedUser: loggedUser, post: post });
+        } catch(err) {
+            console.log(err);
+        }
+    } else {
+        res.send("You are not allowed to access this route");
     }
 
 }
 
 module.exports.post_edit_post = async(req, res) => {
     const { title, body } = req.body;
+    const loggedUser = res.locals.user;
+    const { username } = req.params;
 
-    try {
-        const post = await Post.findOneAndUpdate({ title }, { title, body});
-        res.json({ post });
-    } catch(err){
-        console.log(err);
+    if (loggedUser.username === username) {
+        try {
+            const post = await Post.findOneAndUpdate({ title }, { title, body});
+            res.json({ post });
+        } catch(err){
+            console.log(err);
+        }
+    } else {
+        res.send("You are not allowed to access this route");
     }
 }
 
@@ -83,10 +101,14 @@ module.exports.post_delete = async(req, res) => {
     const { slug } = req.params;
     const { username } = res.locals.user;
 
-    try {
-        await Post.findOneAndDelete({ slug });
-        res.redirect(`/${username}/posts`);
-    } catch(err) {
-        console.log(err);
+    if (loggedUser.username === username) {
+        try {
+            await Post.findOneAndDelete({ slug });
+            res.redirect(`/${username}/posts`);
+        } catch(err) {
+            console.log(err);
+        }
+    } else {
+        res.send("You are not allowed to access this route");
     }
 }
